+++ casual/casual-network/src/main/java/se/kodarkatten/casual/network/messages/transaction/CasualTransactionResourceCommitRequestMessage.java
@@ -13,12 +13,12 @@ import java.util.UUID;
  */
 public final class CasualTransactionResourceCommitRequestMessage extends AbstractCasualTransactionRequestMessage
 {
-    protected CasualTransactionResourceCommitRequestMessage(final UUID execution, final Xid xid, long resourceId, int flags)
+    protected CasualTransactionResourceCommitRequestMessage(final UUID execution, final Xid xid, int resourceId, int flags)
     {
         super(execution, xid, resourceId, flags);
     }
 
-    public static CasualTransactionResourceCommitRequestMessage of(final UUID execution, final Xid xid, long resourceId, final Flag<XAFlags> flags)
+    public static CasualTransactionResourceCommitRequestMessage of(final UUID execution, final Xid xid, int resourceId, final Flag<XAFlags> flags)
     {
         return new CasualTransactionResourceCommitRequestMessage(execution, XID.of(xid), resourceId, flags.getFlagValue());
     }
